Debugged the full backend code.

+++ backend/src/main/java/com/nsbm_projects/hotel_management_system/dto/BookingResponse.java
@@ -4,16 +4,18 @@ import com.nsbm_projects.hotel_management_system.model.BookingStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
 public class BookingResponse {
+
     private Long id;
     private Long guestId;
     private Long roomId;
     private LocalDate checkInDate;
     private LocalDate checkOutDate;
     private BookingStatus bookingStatus;
-    private double totalAmount;
+    private BigDecimal totalAmount;   // ✅ FIXED
 }
