+++ backend/src/main/java/com/nsbm_projects/hotel_management_system/dto/RoomRequest.java
@@ -2,10 +2,13 @@ package com.nsbm_projects.hotel_management_system.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RoomRequest {
+
     private String roomNumber;
     private String type;
-    private double pricePerNight;
+    private BigDecimal pricePerNight;   // ✅ FIXED
     private boolean available;
 }
